A.F....... [ZBX-19248] fixed several issues related to proposed changes

+++ ui/include/classes/api/services/CHostGeneral.php
@@ -172,7 +172,9 @@ abstract class CHostGeneral extends CHostBase {
 				]
 			]);
 			$hostMacroIds = zbx_objectValues($hostMacros, 'hostmacroid');
-			API::UserMacro()->delete($hostMacroIds);
+			if ($hostMacroIds) {
+				API::UserMacro()->delete($hostMacroIds);
+			}
 		}
 
 		if (isset($data['groupids'])) {
